requirements.txt
pywebview==5.3.2
requests==2.32.3
tk==0.1.0